Put LinkPopover class into src/core/gui
