+++ src/core/control/tools/LinkEditor.cpp
@@ -3,75 +3,22 @@
 #include <iostream>
 
 #include <gdk/gdk.h>  // for GdkRectangle, Gdk...
-#include <gtk/gtk.h>  // for GtkWidget, gtk_co...
+#include <gtk/gtk.h>  // for GtkWindow
 
 #include "control/Control.h"           // for Control
 #include "control/zoom/ZoomControl.h"  // for ZoomControl
-#include "gui/GladeSearchpath.h"       // for GladeSearchPath
+#include "gui/LinkPopover.h"           // for LinkPopover
 #include "gui/PageView.h"              // for PageView
 #include "gui/XournalView.h"           // for XournalView
 #include "gui/dialog/LinkDialog.h"     // for LinkDialog
 #include "model/Document.h"            // for Document
+#include "model/Link.h"                // for Link
 #include "model/XojPage.h"             // for XojPage
 #include "undo/InsertUndoAction.h"     // for InsertUndoAction
 #include "util/PopupWindowWrapper.h"   // for PopupWindowWrapper
 #include "util/XojMsgBox.h"            // for XojMsgBox
 #include "util/i18n.h"                 // for FS, _, _F
 
-LinkPopover::LinkPopover(XournalView* view): view(view) {
-    this->popover = GTK_POPOVER(gtk_popover_new(view->getWidget()));
-    gtk_popover_set_modal(popover, false);
-    gtk_popover_set_constrain_to(this->popover, GTK_POPOVER_CONSTRAINT_WINDOW);
-    GtkWidget* vbox = gtk_box_new(GTK_ORIENTATION_VERTICAL, 0);
-    this->label = GTK_LABEL(gtk_label_new(""));
-    gtk_box_pack_start(GTK_BOX(vbox), GTK_WIDGET(this->label), true, true, POPOVER_PADDING);
-    gtk_container_add(GTK_CONTAINER(this->popover), vbox);
-}
-
-LinkPopover::~LinkPopover() {
-    gtk_widget_destroy(GTK_WIDGET(this->label));
-    gtk_widget_destroy(GTK_WIDGET(this->popover));
-}
-
-void LinkPopover::show() { gtk_widget_show(GTK_WIDGET(this->popover)); }
-
-void LinkPopover::show_all() { gtk_widget_show_all(GTK_WIDGET(this->popover)); }
-
-void LinkPopover::hide() { gtk_widget_hide(GTK_WIDGET(this->popover)); }
-
-void LinkPopover::popup() { gtk_popover_popup(this->popover); }
-
-void LinkPopover::popdown() { gtk_popover_popdown(this->popover); }
-
-bool LinkPopover::hasLink() { return (this->link != nullptr); }
-
-void LinkPopover::updateLabel(bool markup) {
-    if (markup) {
-        std::string url = this->link->getUrl();
-        std::string str = "<a href=\"" + url + "\"> " + url + "</a> \n" +
-                          "<span size=\"smaller\"><i> Double click to edit. CTRL + click to open. </i></span>";
-        gtk_label_set_markup(this->label, str.c_str());
-    } else {
-        gtk_label_set_text(this->label, this->link->getUrl().c_str());
-    }
-}
-
-void LinkPopover::positionPopover() {
-    if (this->link) {
-        XojPageView* pageView = view->getViewFor(view->getCurrentPage());
-        auto pos = pageView->getPixelPosition();
-        GdkRectangle rect{pos.x + int(this->link->getX() * pageView->getZoom()),
-                          pos.y + int(this->link->getY() * pageView->getZoom()),
-                          int(this->link->getElementWidth() * pageView->getZoom()),
-                          int(this->link->getElementHeight() * pageView->getZoom())};
-        gtk_popover_set_pointing_to(this->popover, &rect);
-    }
-}
-
-void LinkPopover::linkTo(Link* link) {
-    this->link = link;
-    positionPopover();
-}
 
 LinkEditor::LinkEditor(XournalView* view): view(view), control(view->getControl()) {
     this->highlightPopover = std::make_unique<LinkPopover>(view);
